YG-76 🎁 utils type 분리
