+++ libs/store.ts
@@ -1,19 +1,6 @@
 import create from "zustand"
-import { Dayjs } from "dayjs"
-import { Continent } from "@/constants/continents"
-import { Post } from "@/utils/type"
-import { createPostTemplate, initialFormData } from "@/apis/type"
-
-type SelectionState = {
-    selectedContinent: Continent | null
-    selectedCountry: string | null
-    startDate: Dayjs | null
-    endDate: Dayjs | null
-    setSelectedContinent: (continent: Continent | null) => void
-    setSelectedCountry: (country: string | null) => void
-    setStartDate: (date: Dayjs | null) => void
-    setEndDate: (date: Dayjs | null) => void
-}
+import { initialFormData } from "@/apis/type"
+import { FormState, SelectionState } from "./type"
 
 export const useSelectionStore = create<SelectionState>(set => ({
     selectedContinent: null,
@@ -25,16 +12,6 @@ export const useSelectionStore = create<SelectionState>(set => ({
     setStartDate: date => set({ startDate: date }),
     setEndDate: date => set({ endDate: date }),
 }))
-
-type FormState = {
-    formData: createPostTemplate
-    posts: Post[]
-    quillEditors: Array<{ content: string }>
-    setFormData: (data: createPostTemplate) => void
-    setPosts: (posts: Post[]) => void
-    setQuillEditors: (editors: Array<{ content: string }>) => void
-    resetFormData: () => void
-}
 
 export const useFormDataStore = create<FormState>(set => ({
     formData: initialFormData,
